+++ src/renderer.ts
@@ -19,3 +19,4 @@ export { renderCircleOfInfluence } from "./renderer/circleofinfluence";
 export { renderWholePerson } from "./renderer/wholeperson";
 export { renderRadar } from "./renderer/radar";
 export { renderProblem } from "./renderer/problem";
+export { renderTestCard } from "./renderer/testcard";